% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheSolve.R
\name{cacheSolve}
\alias{cacheSolve}
\title{Returns the inverse of matrix. if cached then return the cached inverse.}
\usage{
cacheSolve(x, ...)
}
\arguments{
\item{x}{- datastructure generated by makeCacheMatrix}
}
\description{
This function computes the inverse of the special "matrix" returned by makeCacheMatrix.
If the inverse has already been calculated (and the matrix has not changed),
then the cachesolve should retrieve the inverse from the cache.
}
\seealso{
\code{\link{solve}} Used to find inverse of a matrix

\code{\link{makeCacheMatrix}} Create necessary data structure for cacheSolve function
}
\author{
Priyesh Kannan
}
